requirements.txt
Django==1.5.4
PIL==1.1.7
South==0.8.2
argparse==1.2.1
dj-database-url==0.2.2
django-braces==1.2.2
django-crispy-forms==1.4.0
django-discover-runner==1.0
django-floppyforms==1.1
django-model-utils==1.5.0
django-registration==1.0
django-widget-tweaks==1.3
psycopg2==2.5.1
six==1.4.1
wsgiref==0.1.2
